# Implement faster gridmake