Minor PR to implement a faster version of `gridmake` as well as implement a mutating version.

Also added some tests.

NOTE: this is breaking because previously we had the first array varying the fastest, now we have the last array varying the fastest. I don't believe this should cause problems for anyone, but if it does let me know and we'll do the other implementation.

- - -

A quick benchmark comparison:

```
julia> a1, a2, a3 = randn(10), randn(10), randn(10);

julia> @benchmark gridmake1(a1, a2, a3)
Benchmarks.ExecutionResults:
  samples:                          (see :samples field)
  precompiled?:                     true
  multiple evaluations per sample?: false
  total time spent benchmarking:    0.48 s
  Benchmarks.SummaryStats:
    estimated time per evaluation: 4.25 ms [3.34 ms, 5.17 ms]
    R² of OLS model:               N/A (OLS not performed)
    estimated % time in GC:        1.96 % [0.0 %, 7.9 %]
    bytes allocated:               819.83 kb
    number of allocations:         25104
    number of samples:             100
    number of evaluations:         100

julia> @benchmark gridmake(a1, a2, a3)
Benchmarks.ExecutionResults:
  samples:                          (see :samples field)
  precompiled?:                     true
  multiple evaluations per sample?: true
  total time spent benchmarking:    10.77 s
  Benchmarks.SummaryStats:
    estimated time per evaluation: 16.13 μs [15.73 μs, 16.52 μs]
    R² of OLS model:               0.91
    estimated % time in GC:        5.52 % [4.3 %, 6.74 %]
    bytes allocated:               26.61 kb
    number of allocations:         71
    number of samples:             6101
    number of evaluations:         670901
```

Here `gridmake` is the new version and `gridmake1` is the old version. This scales better and better the larger the input arrays.